pyproject.toml
[project]
name = "henriquesebastiao"
version = "0.1.0"
description = "Site pessoal e Portifólio"
authors = [
    {name = "Henrique Sebastião",email = "contato@henriquesebastiao.com"}
]
license = {text = "GPL-3"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "rich (>=13.9.4,<14.0.0)",
    "python-decouple (>=3.8,<4.0)"
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint.per-file-ignores]
"check_broken_links.py" = ["E501", "PLR2004"]